pause

+++ app/src/main/java/com/gmail/pentominto/us/supernotes/screens/noteeditscreen/composables/NoteEditScreen.kt
@@ -40,7 +40,7 @@ import kotlinx.coroutines.launch
 
 @Composable
 fun NoteEditScreen(
-    noteId : Long,
+    noteId : Int,
     viewModel : NoteEditScreenViewModel = hiltViewModel()
 
 ) {
